Integer overflow - done

+++ integer_overflow/autograder/sample/student1/starter.sol
@@ -2,22 +2,26 @@ pragma solidity 0.4.8;
 
 contract TotalEarnings {
     
+    
     string answer = "";
-    uint8 revenue;
+    uint256 revenue;
     
     function TotalEarnings() {
         revenue = 0;
     }
     
-    function AddRevenue (uint8 val) public {
+    function AddRevenue (uint256 val) public {
         // Add your code here
         
+        // if (revenue + val > 255) { throw; }
+        // require(revenue + val < 256)
+
         // --
         
         revenue += val;
     }
     
-    function getRevenue() returns (uint8) {
+    function getRevenue() returns (uint256) {
         return revenue;
     }
